Day08/src/ex02/Service/src/main/java/school21/spring/service/services/UsersServiceImpl.java
package school21.spring.service.services;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Component;
import school21.spring.service.models.User;
import school21.spring.service.repositories.UsersRepository;

import java.util.UUID;

@Component("userService")
public class UsersServiceImpl implements UsersService {
    private final UsersRepository repository;

    @Autowired
    public UsersServiceImpl(
            @Qualifier("usersRepositoryJdbcTemplate") UsersRepository repository) {
        this.repository = repository;
    }

    @Override
    public String signUp(String email) {
        Long id = (long) (repository.findAll().size() + 1);
        String password = UUID.randomUUID().toString();
        repository.save(new User(id, email, password));
        return password;
    }
}
